Port 'data.dic' from Spylls
The types are quite small and doable in one shot.

I've ported Flag and FlagSet from prior translations of Nuspell: it
should be possible to support all flag encodings as u32s. Nuspell
actually uses u16s which makes it harder to store UTF8 flags, but is
more efficient in the common case. We can investigate later whether it's
worth the changes necessary to do that. (We would need to port Nuspell's
String_Set which would be nice to avoid.)

I considered SlotMap for Dic.words: using the index of the Vec seems a
little messy compared to SlotMap's nice interface. SlotMap adds overhead
though and only beats a Vec in cases that we don't care about: deleting
keys.