Port lookup functions
This is not fully complete: some types need some members implemented
like CompoundRule. This is the high level code for lookup though and
should be workable for a small dictionary.

The Checker type is custom and removes some default args that are passed
all the way through the lookup functions in spylls, which feels cleaner.