Add a custom CompoundRule type that emulates regex for flags
This is a sort of flag pattern which acts like regex including support
for wildcards, and can also take sets of characters. This matches
the use-case we want for lookup and is less (computationally)
complex than spylls' cartesian product plus regex match.

I have only implemented '?' and '*' for now. Maybe '+' could be
implemented in the future but I don't see any dictionaries that use it.