Add parsers for .dic and .aff formats
This is mostly based on zspell's parser which is parser-combinator-like
(sort of, vaguely). Also included is an overhaul of many of the types
to make them easier to parse, easier to consume, or for correctness.

This does not fully parse the Aff: some commands like ICONV and MAP
are left for future work.