Lookup homonyms for each generated affix form
The prior `word` value was a typo: we need to check the homonyms for
the generated affix form's stem. For example if we are looking up
"retry" in en_US, `produce_affix_forms` will desuffix the word into
"retry" and "try." Previously we only checked the word ("retry") but
we should be checking the stem of those forms ("try"). This fixes
lookup for words which are only in the dictionary via affix rules.