Reverse affix chars when inserting suffixes into the index
We lookup and enter the 'affix'/'add' part of the suffix reversed so we
need to enter it reversed as well. This is just a copy/paste typo from
prefixes which we enter from the start of the stem/prefix/aff text.

This fixes lookup for any word created through suffixes. An easy example
is "it" looked up well before this commit but not "its."