Rewrite AffixIndex to allow returning all values along the keypath
So that a word like "complications" will return suffixes for "s",
"ions", and "ications."

This fixes any non-trivial suffix lookup ("its" worked previously but
not "scions"). It's similar to Nuspell's Prefix/Suffix_Multiset, though
a different approach as far as I can read of the C++.

The `Get` type here might also be good inspiration for how we can
convert a lot of the `-> Vec<T>` returns we have in `Checker` to
iterators, so we can reduce allocations. I think it would muddle the
readability to convert everything to iterators but it may be worth it
performance-wise.