Check affix flags
This is an oversight from porting spylls' `is_good_form` method. We need
to check that a prefix is allowed to be used on a stem by comparing the
Word's flag set against the affix's flag. This fixes lookups for
incorrect words like "undrink". It's a valid construction of a prefix
("un") with a word from the dic ("drink") but the "un" prefix is not
in "drink"'s flags, so it should be rejected.