AffixForm: Remove unnecessary 'in_dictionary' field
This field is a bit awkward: it's None most of the time until we pass
the AffixForm into 'is_good_form'. We can instead just pass a borrow of
the Word to 'is_good_form' which also eliminates some eager clones.

I see a ~8-15% performance boost for this in the benchmark.