Vendor the en_US dictionary
It will be a little friendlier for anyone trying out the examples or
benchmarks to make the dictionary lookup infallible. We will eventually
want to use a real dictionary for integration tests and we can drop the
xdg dependency. The file size is around a half of a MB so I don't feel
too badly about pulling this in.