Allow passing in the BuildHasher for WordList
This should allow us to drop the dependency on ahash and only depend on
it as a dev-dependency. Downstream consumers can then pass in whichever
hasher they would like.

This was difficult in the first attempt at spellbook because of parsing:
the strategy of applying a set of parsing functions for each directive
doesn't work when the functions have to be generic for the BuildHasher.
In the child commit we will split the AffData type into its options and
its non-trivial data structures which will make this possible.

We could consider also moving the WordList to its own member on
Dictionary so that AffData would not need to be generic at all. It's
nice to have AffData's options split out though because we can trivially
parse that data while most of the data structures benefit from first
being parsed into a Vec or other basic collection and then converted
into the end data structure (for example AffixIndex needs this - adding
elements piece-wise is not efficient.)