Implement compound_rule_matches and CompoundRuleTable::any_rule_matches
The compound rule matching is slightly complicated because of the
wildcards. Nuspell approaches this with a stack vector rather than
recursion.

I'm not sure the difference would ever be noticeable between either
the stack vs. recursion or our differences in representation. Poking
around JetBrains/hunspell-dictionaries, it seems like compound rules
are used so rarely and so little that it should be negligible.