Add the very basics of checking un-cased words
This is a port of the casing enum and detection function from Nuspell
along with some of the most basic checking functionality.

I've also set up the tests in the checker to be able to take advantage
of the vendored en_US dictionary. This does make the tests quite a bit
slower than they were (basically instant) but it's still comfortably
sub-second.