Replace `is_some_and` with has_flag! macro for common case
We only use is_some_and so far to check if one of the optional flags on
the AffOptions is present in a FlagSet. is_some_and is a little verbose
so we'll replace it with a macro that covers the match. Hopefully this
should be as efficient as inlining the is_some_and function.