Store the word list stem on CompoundingResult & AffixForm
This approximately matches Nuspell's storage of the wordlist pointer.
To make the lifetimes play nice we return the key from the map as well
as the value. This value is equal to the `stem2`/`stem3`s in the
affixing functions but we need to return the data from the map to use
that liftime rather than the Cows produced by stripping and adding with
affixes.

As another happy consequence of this, we can drop the `'aff` lifetime
on the `word: &str` parameters in the affixing stripping functions. The
input word should have a distinct lifetime and not having this lifetime
will cause problems for the compounding functions introduced in the
child commits.