Implement check for COMPOUNDRULE compounds
This fixes check for compounds in en_US like "10th" and "202nd". These
are considered valid because of the COMPOUNDRULE rule. From en_US.aff:

    # compound rules:
    # 1. [0-9]*1[0-9]th (10th, 11th, 12th, 56714th, etc.)
    # 2. [0-9]*[02-9](1st|2nd|3rd|[4-9]th) (21st, 22nd, 123rd, 1234th, etc.)
    COMPOUNDRULE 2
    COMPOUNDRULE n*1t
    COMPOUNDRULE n*mp

For example with dictionary entries "1/n1" and "0th/pt", we split up
the word "10th" into parts "1" and "0th" which we look up in the
dictionary. We then check those words flagsets against the patterns.
"10th" would be represented as:

    `&[&flagset!['n', '1'], &flagset!['p', 't']]`

That matches the pattern `n*1t`: zero or more 'n' flags and then a '1'
flag (used for "1" and other digits) and a 't' flag ("0th"). The 'n'
flag matches zero times for "10th" but allows other numbers in front.
For example this pattern would also match "110th".

These kinds of compounds are fairly straightforward to check compared
to the other kind of compounding. This is the only compounding used by
en_US and a few other dictionaries.

This completes support for en_US for the checker - nothing else in the
aff affects the checker.