Perform case conversion based on locale
A few Turkic locales have special conversion rules for 'i' and 'I' which
we need to handle.

Nuspell covers this with ICU but we don't want to pull in ICU4X - it's
a very large dependency and adds weight even if we just pull in the
case mapping parts. Luckily the code to do this ourselves (like
Hunspell does too) is very small.