Add a blank impl of `check_compound_with_pattern_replacements`
This function only does something when CompoundResult has a replacement
field on it. In my searching I didn't see a dictionary that actually
uses that so we might want to hold off on implementing that function
(which is really very complex in the Nuspell codebase, with lots of
`goto`) until we find a language that uses this.