aff parser: Only use dictionary number as a capacity hint
This seems to match what Nuspell does - it uses the number at the
beginning of the dic file as an approximate size that it pre-allocates
within its word list.

We already used this capacity hint but we stopped parsing after reading
that many rows. This caused some failures in the Hunspell corpus where
Nuspell added words to the example dictionaries but didn't update the
numbers.