checker: Draft out classic compound checking
With this change the checker should only be missing classic compound
checking with CHECKCOMPOUNDPATTERN rules with replacements. That
function is similarly as complex as this one so I'd like to avoid
implementing that for now, at least until we can find a dictionary that
actually uses them.

This was tricky to translate from Nuspell since it uses `goto` to do the
different checks. I'm sure I have some mistakes in here somewhere.