checker: Fix cross-product check in commutative affixing
Fixes the check for "undrinks". It was a copy/paste error from the above
definition of `valid_cross_prefix_outer`: we need to check that the
wordlist entry's flags allow the given prefix ("un" in this case). We
accidentally checked that the suffix ("s" in this case) was allowed
twice.

This fixes the "conditionalprefix" test case from the Hunspell corpus
which specifically checked "undrinks."