parser: Improve handling of spaces in .dic file words
This fixes the `hu` case from the corpus. There's work to be done
though: `parse_dic_line` needs to be rewritten and follow Nuspell. It's
a complicated parser to write though. This commit passes the entire
line from the .dic file and lets `parse_dic_line` handle it. There's a
commented out case in the `parse_dic_line_test` unit test that should
also be handled but isn't currently tested by the corpus.