parser: Rewrite `parse_dic_line` to properly handle morphs and spaces
This is the continuation of a parent commit (60aec39f5494) after I've
slept and had lots of coffee. We want to preserve as many fast-lanes as
we can for the common cases: a word not having a flagset and a word only
having a flagset (i.e. not using spaces or backslash escaping). Then if
we detect a word uses spaces or backslashes we switch to building the
word gradually with a char iterator.

This is a bit tricky to translate from Nuspell because it has the input
line allocated into a string. That allows it to handle backslashes
pretty easily: it erases the backslash character and skips over the
slash in a loop that finds the end of the word part of the input. Then
it handles flagsets.

I believe what we have in this commit is potentially faster than what
Nuspell does since we avoid allocation in the common cases. I expected
a regression in performance from the prior implementation before the
change a few commits back but they appear almost identical for en_US at
least. (In retrospect this is not totally surprising: en_US should only
ever hit the fast lanes.)