parser: Fix "slash" case from corpus
This is a little bit of a bandaid fix. I believe what Nuspell is doing
is allowing '/' at the start of a word. These comment-like entries like
in it_IT actually end up in the word list but are never hit in practice
because of tokenization. Here I'm choosing to omit them from the word
list instead. I think this shouldn't really matter in practice since the
tokenizer should probably not allow '/' at the beginning of words
anyways.