checker: Fix ONLYINCOMPOUND flagset check
We need to check the flagset of the word in the wordlist rather than
the prefix/suffix being considered. A stem might be marked as
only-in-compound (as well as a prefix or suffix in its continuation
flags). This follows Nuspell and may have been a bad yank/paste error
on my behalf.