ICONV/OCONV: Favor the earliest, longest replacement pattern
This fixes the nepali case from the corpus (alongside with the parent
commit which added support for '_' anchors in ICONV/OCONV rules).

To fix this we need to fix the lookup of conversion patterns to favor
the one that comes earliest in the input word and also has the longest
length of any that convert at that byte.