Move macros into their calling modules
This causes some duplication but these macros are very simple and short.
The issue is that we can't describe a macro as available throughout a
crate but not exported from that crate. Basically what we want is a
`pub(crate)` macro, and that doesn't seem to be possible in the current
Rust edition. If that's fixed in the future we could reintroduce this
macros module and move all of the macros there.

This removes the macros from the `cargo doc` output notably.