suggester: Deduplicate suggestions
This is done in Nuspell just before OCONV and it prevents suggestions
for "adveenture" from producing "adventure" twice. (Which might happen
since the `extra_char_suggest` procedure will eliminate both innner 'e'
characters.)

Not sure if ordering is important but this should help with Hunspell
compatibility since it seems to have the same suggestions and ordering
as Nuspell.