flake: Add valgrind to the devShell
valgrind is useful for the DHAT heap analysis tool - allowing you to
inspect heap interaction at runtime. For example you can compile one
of the example programs:

    cargo run --release --example check hello

and then run it while analyzing the heap interaction:

    valgrind --tool=dhat ./target/release/examples/check hello

And the result is a report on the heap usage. You can load it into
your browser to poke around the allocations and see what specifically
is taking a lot of memory.