suggester: Refactor out unnecessary `unsafe` string modification
At the end of the day the old and new code are equivalent.
`String::insert` does slightly more work (calculating the UTF-8 length
and such) which is unnecessary since we know that `' '.len_utf8() == 1`,
but this string modification is done at most once per suggestions so the
savings are not perceptible.