Switch default hasher feature to use `foldhash` instead of `ahash`
This was intended to follow `hashbrown` v0.15's switch. There is also a
nice side effect of dropping proc-macro stuff that I didn't realize that
ahash had introduced, making the `cargo tree` and `Cargo.lock`
delightfully small.

As previously remarked, the switch does not seem to change performance
either for the better or worse.