try just using a pointer
Otherwise the Arc needs to hold another usize to tell the number of
entries, so it knows how to deallocate the pointer. We can do this
ourselves so that the size of a `Arc<Subtrie<()>>` is 8 (pointer size)
instead of 16. The arc storing the length is redundant since it's
already stored in the bitmap.