Default `forbidden_word_flag` to `Flag(65510)`
Hunspell defaults this flag to 65510 so that words can be marked as
forbidden even if the dictionary does not set a `FORBIDDENWORD` flag
explicitly. So instead of an `Option<Flag>` this is a `Flag`.

This includes a logic change in the suggester to eliminate a check that
the forbidden_word_flag is Some. Because of the non-zero default in
Hunspell this branch is always `true` and can be removed.